Rename initalRatings to correct lower case name
